feat: add solutions to lc problem: No.0699 (#3328)
No.0699.Falling Squares

+++ solution/0600-0699/0699.Falling Squares/Solution.go
@@ -9,7 +9,7 @@ func newNode(l, r int) *node {
 	return &node{
 		l:   l,
 		r:   r,
-		mid: int(uint(l+r) >> 1),
+		mid: (l + r) >> 1,
 	}
 }
